multi_agent_system/requirements.txt
# Multi-Agent Hospital Operations System Requirements

# Core Framework
langgraph>=0.5.3
langchain>=0.3.26
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Logging & Monitoring
structlog>=23.0.0

# Testing & Development
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utility
uuid7>=0.1.0
python-dateutil>=2.8.0
